+++ test/unit/test_C_detector.py
@@ -7,7 +7,7 @@ import pytest
 from pii_data.helper.exception import ConfigException
 from pii_extract.gather.collection import get_task_collection
 
-from taux.monkey_patch import patch_entry_points, patch_transformer_pipeline
+from taux.monkey_patch import patch_entry_points, patch_transformer_pipeline, patch_env
 
 # ---------------------------------------------------------------------------
 
@@ -18,12 +18,13 @@ def test10_tasklist(monkeypatch):
     """
     patch_entry_points(monkeypatch)
     patch_transformer_pipeline(monkeypatch, {}, model_labels=["PER", "LOC"])
+    patch_env(monkeypatch)
 
     piic = get_task_collection(debug=False)
     tasks = piic.build_tasks()
     tasks = list(tasks)
     assert len(tasks) == 1
-    assert str(tasks[0]) == "<Transformers wrapper #4>"
+    assert str(tasks[0]) == "<Transformers wrapper #6>"
 
 
 def test11_tasklist_err(monkeypatch):
@@ -32,6 +33,7 @@ def test11_tasklist_err(monkeypatch):
     """
     patch_entry_points(monkeypatch)
     patch_transformer_pipeline(monkeypatch, {})
+    patch_env(monkeypatch)
 
     piic = get_task_collection(debug=False)
     tasks = piic.build_tasks()
@@ -45,6 +47,7 @@ def test20_tasklist_lang(monkeypatch):
     """
     patch_entry_points(monkeypatch)
     patch_transformer_pipeline(monkeypatch, {}, model_labels=["PER", "LOC"])
+    patch_env(monkeypatch)
 
     piic = get_task_collection(debug=False)
     tasks = piic.build_tasks('en')
@@ -58,6 +61,8 @@ def test30_engine_cache(monkeypatch):
     Check engine object reuse
     """
     patch_entry_points(monkeypatch)
+    patch_env(monkeypatch)
+
     mck = patch_transformer_pipeline(monkeypatch, {}, model_labels=["PER", "LOC"])
 
     piic = get_task_collection(debug=True)